requirements.txt
xgboost==1.7.6
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
shap==0.42.1
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
category-encoders==2.6.2
